+++ tests/testthat/test-005-pro_fetch_search.R
@@ -20,7 +20,6 @@ test_that("pro_fetch search `biodiversity AND fiance`", {
     pro_fetch(
       pages = 1,
       project_folder = output_dir,
-      mailto = "test@example.com",
       verbose = FALSE,
       progress = TRUE
     )