+++ tests/testthat/test-007-pro_pipeline_group_by.R
@@ -23,7 +23,6 @@ test_that("pro_request `biodiversity` and group by `type`", {
     pro_request(
       pages = 1,
       output = output_json,
-      mailto = "test@example.com",
       verbose = FALSE,
       progress = TRUE
     )